Wrap up http request and response debugging properly in a net/http round tripper transport.
