+++ idx.go
@@ -26,11 +26,8 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
-	"net/http/httputil"
 	"net/url"
-	"os"
 	"runtime"
 	"strconv"
 	"strings"
@@ -63,7 +60,6 @@ type AccessToken struct {
 type Client struct {
 	config     *Config
 	httpClient *http.Client
-	debug      bool
 }
 
 type Context struct {
@@ -112,11 +108,11 @@ func NewClientWithSettings(conf ...ConfigSetter) (*Client, error) {
 		return nil, fmt.Errorf("invalid configuration: %w", err)
 	}
 	c := &Client{
-		config:     cfg,
-		httpClient: &http.Client{Timeout: defaultTimeout},
-	}
-	if os.Getenv("DEBUG_IDX_CLIENT") != "" {
-		c.debug = true
+		config: cfg,
+		httpClient: &http.Client{
+			Transport: newIdxTransport(),
+			Timeout:   defaultTimeout,
+		},
 	}
 
 	idx = c
@@ -156,7 +152,7 @@ func (c *Client) introspect(ctx context.Context, ih *InteractionHandle) (*Respon
 	req.Header.Add("Content-Type", "application/ion+json; okta-version=1.0.0")
 	req.Header.Add("Accept", "application/ion+json; okta-version=1.0.0")
 	withOktaUserAgent(req)
-	resp, err := c.httpClientDo(req)
+	resp, err := c.httpClient.Do(req)
 	if err != nil {
 		return nil, fmt.Errorf("http call has failed: %w", err)
 	}
@@ -201,7 +197,7 @@ func (c *Client) interact(ctx context.Context, opts *interactOptions) (*Context,
 	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
 	withDeviceContext(ctx, req)
 
-	resp, err := c.httpClientDo(req)
+	resp, err := c.httpClient.Do(req)
 	if err != nil {
 		return nil, fmt.Errorf("http call has failed: %w", err)
 	}
@@ -254,7 +250,7 @@ func (c *Client) RedeemInteractionCode(ctx context.Context, idxContext *Context,
 	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
 	withOktaUserAgent(req)
 
-	resp, err := c.httpClientDo(req)
+	resp, err := c.httpClient.Do(req)
 	if err != nil {
 		return nil, fmt.Errorf("error calling token api: %w", err)
 	}
@@ -570,63 +566,3 @@ func securityQuestionOptions(ctx context.Context, idxContext *Context) (*Respons
 	m["custom"] = "Create a security question"
 	return resp, m, nil
 }
-
-func (c *Client) debugRequest(req *http.Request) {
-	if req == nil {
-		return
-	}
-	reqData, err := httputil.DumpRequest(req, true)
-	if err == nil {
-		log.Printf("[DEBUG] "+logReqMsg, req.RequestURI, prettyPrintJsonLines(reqData))
-	} else {
-		log.Printf("[ERROR] %s API Request error: %#v", req.RequestURI, err)
-	}
-}
-
-func (c *Client) debugResponse(resp *http.Response) {
-	if resp == nil {
-		return
-	}
-	respData, err := httputil.DumpResponse(resp, true)
-	if err == nil {
-		log.Printf("[DEBUG] "+logRespMsg, resp.Request.RequestURI, prettyPrintJsonLines(respData))
-	} else {
-		log.Printf("[ERROR] %s API Response error: %#v", resp.Request.RequestURI, err)
-	}
-}
-
-func (c *Client) httpClientDo(req *http.Request) (*http.Response, error) {
-	if c.debug {
-		c.debugRequest(req)
-	}
-	resp, err := c.httpClient.Do(req)
-	if c.debug {
-		c.debugResponse(resp)
-	}
-
-	return resp, err
-}
-
-// prettyPrintJsonLines iterates through a []byte line-by-line, transforming any
-// lines that are complete json into pretty-printed json.
-func prettyPrintJsonLines(b []byte) string {
-	parts := strings.Split(string(b), "\n")
-	for i, p := range parts {
-		if b := []byte(p); json.Valid(b) {
-			var out bytes.Buffer
-			json.Indent(&out, b, "", " ")
-			parts[i] = out.String()
-		}
-	}
-	return strings.Join(parts, "\n")
-}
-
-const logReqMsg = `%s API Request Details:
----[ REQUEST ]---------------------------------------
-%s
------------------------------------------------------`
-
-const logRespMsg = `%s API Response Details:
----[ RESPONSE ]--------------------------------------
-%s
------------------------------------------------------`
