+++ response.go
@@ -123,7 +123,7 @@ func (r *Response) Cancel(ctx context.Context) (*Response, error) {
 	req.Header.Set("Accepts", r.CancelResponse.Accepts)
 	req.Header.Set("Content-Type", r.CancelResponse.Accepts)
 	withOktaUserAgent(req)
-	resp, err := idx.httpClientDo(req)
+	resp, err := idx.httpClient.Do(req)
 	if err != nil {
 		return nil, fmt.Errorf("http call has failed: %w", err)
 	}
